Fix timer delay commit and add snap build diagnostics

+++ lib/process.h
@@ -34,9 +34,9 @@
 /* The maximum pid is 2^22 - see definition of PID_MAX_LIMIT in kernel source include/linux/threads.h */
 #define PID_MAX_DIGITS		7
 
-extern unsigned min_auto_priority_delay;
+extern long min_auto_priority_delay;
 
-extern void set_process_priorities(int, int, unsigned,
+extern void set_process_priorities(int, int, long,
 #if HAVE_DECL_RLIMIT_RTTIME == 1
 			           int,
 #endif
